Add support for pyolite
